pyproject.toml

[project]
name = "my-claude-ansible"
version = "0.1.0"
description = "Claude Code 配置管理仓库 - 基于 Ansible 的声明式配置管理"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ansible>=11.5.0",
    "check-jsonschema>=0.30.0",
    "typer>=0.12.0",
    "ruamel.yaml>=0.18.0",
    "rich>=13.0.0",
]

[project.scripts]
my-claude = "my_claude.cli:app"

# Python 换源
[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple"
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["filter_plugins/*", "my_claude"]